created a new boxer test file
